pyproject.toml
[project]
name = "foundry-data-analyst"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "azure-ai-projects>=1.0.0b12",
    "azure-identity>=1.23.0",
    "azure-storage-blob>=12.25.1",
    "python-dotenv>=1.1.1",
]
